+++ cli.js
@@ -3,9 +3,8 @@
 const program = require('commander')
 const pkg = require('./package.json')
 const {fetchValues, isHealthy, printValues} = require('./utils')
-const projectId = 'ntnu-smartmedia'
 
-async function healthCheck (subscriptionId, sampleMinutes, confidence) {
+async function healthCheck (subscriptionId, sampleMinutes, confidence, projectId) {
   const values = await fetchValues(projectId, subscriptionId, sampleMinutes)
   console.log('Calculating subscription health for ' + subscriptionId)
   printValues(values)
@@ -13,6 +12,8 @@ async function healthCheck (subscriptionId, sampleMinutes, confidence) {
   console.log('Healthy?', healthy ? 'YES' : 'NO')
   process.exit(healthy ? 0 : 1)
 }
+
+let taken = false
 
 program
   .version(pkg.version)
@@ -22,8 +23,23 @@ program
   .description('Check pubsub subscription health')
   .option('-m --sampleMinutes <minutes>', 'Sample duration in minutes. Default is 20 minutes.', parseInt)
   .option('-c --confidence <confidence>', 'Confidence for passing between 0 and 1. Default is 0.75', parseFloat)
+  .option('-p --project <project>', 'Google Cloud project name. Defaults to "ntnu-smartmedia"')
   .action((subscriptionId, options) => {
-    healthCheck(subscriptionId, options.sampleMinutes || 20, options.confidence || 0.75)
+    taken = true
+    healthCheck(
+      subscriptionId,
+      options.sampleMinutes || 20,
+      options.confidence || 0.75,
+      options.project || 'ntnu-smartmedia')
+      .catch(err => {
+        console.error('Failed with error:', err)
+        process.exit(2)
+      })
   })
 
 program.parse(process.argv)
+
+if (!taken) {
+  program.outputHelp()
+  process.exit(2)
+}