+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Add project option
+- Add more health messages
+- Updated documentation
+
 ## [0.0.0] 
 
  - Initial commit